+++ apps/minifront/src/components/swap/swap-form/token-swap-input.tsx
@@ -3,13 +3,8 @@ import { BalancesResponse } from '@penumbra-zone/protobuf/penumbra/view/v1/view_
 import { Metadata } from '@penumbra-zone/protobuf/penumbra/core/asset/v1/asset_pb';
 import { Box } from '@penumbra-zone/ui/components/ui/box';
 import { joinLoHiAmount } from '@penumbra-zone/types/amount';
-import {
-  getAmount,
-  getBalanceView,
-  getMetadataFromBalancesResponse,
-} from '@penumbra-zone/getters/balances-response';
+import { getAmount, getBalanceView } from '@penumbra-zone/getters/balances-response';
 import { ArrowRight } from 'lucide-react';
-import { useMemo } from 'react';
 import { AllSlices } from '../../../state';
 import { useStoreShallow } from '../../../utils/use-store-shallow';
 import { getFormattedAmtFromValueView } from '@penumbra-zone/types/value-view';
@@ -21,13 +16,12 @@ import { isValidAmount } from '../../../state/helpers';
 import { NonNativeFeeWarning } from '../../shared/non-native-fee-warning';
 import { NumberInput } from '../../shared/number-input';
 import { useBalancesResponses, useAssets } from '../../../state/shared';
-import { FadeIn } from '@penumbra-zone/ui/components/ui/fade-in';
 import { getBalanceByMatchingMetadataAndAddressIndex } from '../../../state/swap/getters';
 import {
   swappableAssetsSelector,
   swappableBalancesResponsesSelector,
 } from '../../../state/swap/helpers';
-import { getDisplayDenomExponent } from '@penumbra-zone/getters/metadata';
+import { TokenInputError } from './token-input-error.tsx';
 
 const getAssetOutBalance = (
   balancesResponses: BalancesResponse[] = [],
@@ -69,9 +63,6 @@ export const TokenSwapInput = () => {
   const { amount, setAmount, assetIn, setAssetIn, assetOut, setAssetOut, reverse } =
     useStoreShallow(tokenSwapInputSelector);
   const assetOutBalance = getAssetOutBalance(balancesResponses?.data, assetIn, assetOut);
-  const assetInExponent = useMemo(() => {
-    return getDisplayDenomExponent.optional(getMetadataFromBalancesResponse.optional(assetIn));
-  }, [assetIn]);
 
   const maxAmount = getAmount.optional(assetIn);
   const maxAmountAsString = maxAmount ? joinLoHiAmount(maxAmount).toString() : undefined;
@@ -84,7 +75,7 @@ export const TokenSwapInput = () => {
   };
 
   return (
-    <Box label='Trade' layout>
+    <Box label='Trade' layout headerContent={<TokenInputError />}>
       <div className='flex flex-col items-stretch gap-4 sm:flex-row'>
         <NumberInput
           value={amount}
@@ -92,7 +83,6 @@ export const TokenSwapInput = () => {
           variant='transparent'
           placeholder='Enter an amount...'
           max={maxAmountAsString}
-          maxExponent={assetInExponent}
           step='any'
           className={'font-bold leading-10 md:h-8 md:text-xl xl:h-10 xl:text-3xl'}
           onChange={e => {
@@ -107,13 +97,6 @@ export const TokenSwapInput = () => {
               </span>
             </div>
           )}
-
-          <FadeIn condition={!!balancesResponses?.error || !!swappableAssets?.error}>
-            <div className='flex gap-4 text-red'>
-              {balancesResponses?.error instanceof Error && balancesResponses.error.toString()}
-              {swappableAssets?.error instanceof Error && swappableAssets.error.toString()}
-            </div>
-          </FadeIn>
 
           <div className='flex gap-4'>
             <div className='flex h-full flex-col justify-end gap-2'>